requirements.txt
alembic==0.9.4
aniso8601==1.2.1
click==6.7
coverage==4.4.1
flasgger==0.6.6
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.1.0
Flask-RESTful==0.3.6
flask-restful-swagger==0.19
flask-restful-swagger-2==0.35
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mistune==0.7.4
nose==1.3.7
psycopg2==2.7.3
PyJWT==1.5.2
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
six==1.10.0
SQLAlchemy==1.1.12
Werkzeug==0.12.2
